Move random generations to ray/common/test_util.h
