+++ src/ray/common/id.h
@@ -531,13 +531,5 @@ DEFINE_UNIQUE_ID(ObjectID);
 #include "id_def.h"
 
 #undef DEFINE_UNIQUE_ID
-
 }  // namespace std
-
-namespace ray {
-
-TaskID RandomTaskId();
-
-}  // namespace ray
-
 #endif  // RAY_ID_H_
