+++ src/ray/common/test_util.cc
@@ -1,7 +1,9 @@
-#include "ray/util/test_util.h"
+#include "ray/common/test_util.h"
 
 #include <functional>
 
+#include "ray/common/buffer.h"
+#include "ray/common/ray_object.h"
 #include "ray/util/logging.h"
 
 namespace ray {
@@ -45,6 +47,28 @@ bool WaitForCondition(std::function<bool()> condition, int timeout_ms) {
     }
   }
   return false;
+}
+
+TaskID RandomTaskId() {
+  std::string data(TaskID::Size(), 0);
+  FillRandom(&data);
+  return TaskID::FromBinary(data);
+}
+
+std::shared_ptr<Buffer> GenerateRandomBuffer() {
+  auto seed = std::chrono::high_resolution_clock::now().time_since_epoch().count();
+  std::mt19937 gen(seed);
+  std::uniform_int_distribution<> dis(1, 10);
+  std::uniform_int_distribution<> value_dis(1, 255);
+
+  std::vector<uint8_t> arg1(dis(gen), value_dis(gen));
+  return std::make_shared<LocalMemoryBuffer>(arg1.data(), arg1.size(), true);
+}
+
+std::shared_ptr<RayObject> GenerateRandomObject(
+    const std::vector<ObjectID> &inlined_ids) {
+  return std::shared_ptr<RayObject>(
+      new RayObject(GenerateRandomBuffer(), nullptr, inlined_ids));
 }
 
 /// Path to redis server executable binary.
