+++ src/ray/core_worker/test/mock_worker.cc
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 #define BOOST_BIND_NO_PLACEHOLDERS
+#include "ray/common/test_util.h"
 #include "ray/core_worker/context.h"
 #include "ray/core_worker/core_worker.h"
-#include "src/ray/util/test_util.h"
 
 using namespace std::placeholders;
