+++ src/ray/gcs/test/accessor_test_base.h
@@ -20,10 +20,11 @@
 #include <string>
 #include <thread>
 #include <vector>
+
 #include "gtest/gtest.h"
+#include "ray/common/test_util.h"
 #include "ray/gcs/redis_accessor.h"
 #include "ray/gcs/redis_gcs_client.h"
-#include "ray/util/test_util.h"
 
 namespace ray {
