+++ src/ray/gcs/test/redis_actor_info_accessor_test.cc
@@ -17,10 +17,11 @@
 #include <string>
 #include <thread>
 #include <vector>
+
 #include "gtest/gtest.h"
+#include "ray/common/test_util.h"
 #include "ray/gcs/redis_gcs_client.h"
 #include "ray/gcs/test/accessor_test_base.h"
-#include "ray/util/test_util.h"
 
 namespace ray {
