+++ src/ray/gcs/test/redis_job_info_accessor_test.cc
@@ -13,11 +13,12 @@
 // limitations under the License.
 
 #include <memory>
+
 #include "gtest/gtest.h"
+#include "ray/common/test_util.h"
 #include "ray/gcs/pb_util.h"
 #include "ray/gcs/redis_gcs_client.h"
 #include "ray/gcs/test/accessor_test_base.h"
-#include "ray/util/test_util.h"
 
 namespace ray {
